+++ test/test_external.py
@@ -43,8 +43,8 @@ class TestLoadInterface:
             alias=model_type,
         )
         assert interface.__name__ == model_type
-        assert isinstance(interface.input_components[0], gr.components.Audio)
-        assert isinstance(interface.output_components[0], gr.components.Audio)
+        assert isinstance(interface.input_components[0], gr.Audio)
+        assert isinstance(interface.output_components[0], gr.Audio)
 
     def test_question_answering(self):
         model_type = "image-classification"
@@ -54,8 +54,8 @@ class TestLoadInterface:
             alias=model_type,
         )
         assert interface.__name__ == model_type
-        assert isinstance(interface.input_components[0], gr.components.Image)
-        assert isinstance(interface.output_components[0], gr.components.Label)
+        assert isinstance(interface.input_components[0], gr.Image)
+        assert isinstance(interface.output_components[0], gr.Label)
 
     def test_text_generation(self):
         model_type = "text_generation"
@@ -63,8 +63,8 @@ class TestLoadInterface:
             "models/gpt2", alias=model_type, description="This is a test description"
         )
         assert interface.__name__ == model_type
-        assert isinstance(interface.input_components[0], gr.components.Textbox)
-        assert isinstance(interface.output_components[0], gr.components.Textbox)
+        assert isinstance(interface.input_components[0], gr.Textbox)
+        assert isinstance(interface.output_components[0], gr.Textbox)
         assert any(
             "This is a test description" in d["props"].get("value", "")
             for d in interface.get_config_file()["components"]
@@ -76,8 +76,8 @@ class TestLoadInterface:
             "models/facebook/bart-large-cnn", api_key=None, alias=model_type
         )
         assert interface.__name__ == model_type
-        assert isinstance(interface.input_components[0], gr.components.Textbox)
-        assert isinstance(interface.output_components[0], gr.components.Textbox)
+        assert isinstance(interface.input_components[0], gr.Textbox)
+        assert isinstance(interface.output_components[0], gr.Textbox)
 
     def test_translation(self):
         model_type = "translation"
@@ -85,8 +85,8 @@ class TestLoadInterface:
             "models/facebook/bart-large-cnn", api_key=None, alias=model_type
         )
         assert interface.__name__ == model_type
-        assert isinstance(interface.input_components[0], gr.components.Textbox)
-        assert isinstance(interface.output_components[0], gr.components.Textbox)
+        assert isinstance(interface.input_components[0], gr.Textbox)
+        assert isinstance(interface.output_components[0], gr.Textbox)
 
     def test_text2text_generation(self):
         model_type = "text2text-generation"
@@ -94,8 +94,8 @@ class TestLoadInterface:
             "models/sshleifer/tiny-mbart", api_key=None, alias=model_type
         )
         assert interface.__name__ == model_type
-        assert isinstance(interface.input_components[0], gr.components.Textbox)
-        assert isinstance(interface.output_components[0], gr.components.Textbox)
+        assert isinstance(interface.input_components[0], gr.Textbox)
+        assert isinstance(interface.output_components[0], gr.Textbox)
 
     def test_text_classification(self):
         model_type = "text-classification"
@@ -105,8 +105,8 @@ class TestLoadInterface:
             alias=model_type,
         )
         assert interface.__name__ == model_type
-        assert isinstance(interface.input_components[0], gr.components.Textbox)
-        assert isinstance(interface.output_components[0], gr.components.Label)
+        assert isinstance(interface.input_components[0], gr.Textbox)
+        assert isinstance(interface.output_components[0], gr.Label)
 
     def test_fill_mask(self):
         model_type = "fill-mask"
@@ -114,8 +114,8 @@ class TestLoadInterface:
             "models/bert-base-uncased", api_key=None, alias=model_type
         )
         assert interface.__name__ == model_type
-        assert isinstance(interface.input_components[0], gr.components.Textbox)
-        assert isinstance(interface.output_components[0], gr.components.Label)
+        assert isinstance(interface.input_components[0], gr.Textbox)
+        assert isinstance(interface.output_components[0], gr.Label)
 
     def test_zero_shot_classification(self):
         model_type = "zero-shot-classification"
@@ -123,10 +123,10 @@ class TestLoadInterface:
             "models/facebook/bart-large-mnli", api_key=None, alias=model_type
         )
         assert interface.__name__ == model_type
-        assert isinstance(interface.input_components[0], gr.components.Textbox)
-        assert isinstance(interface.input_components[1], gr.components.Textbox)
-        assert isinstance(interface.input_components[2], gr.components.Checkbox)
-        assert isinstance(interface.output_components[0], gr.components.Label)
+        assert isinstance(interface.input_components[0], gr.Textbox)
+        assert isinstance(interface.input_components[1], gr.Textbox)
+        assert isinstance(interface.input_components[2], gr.Checkbox)
+        assert isinstance(interface.output_components[0], gr.Label)
 
     def test_automatic_speech_recognition(self):
         model_type = "automatic-speech-recognition"
@@ -134,8 +134,8 @@ class TestLoadInterface:
             "models/facebook/wav2vec2-base-960h", api_key=None, alias=model_type
         )
         assert interface.__name__ == model_type
-        assert isinstance(interface.input_components[0], gr.components.Audio)
-        assert isinstance(interface.output_components[0], gr.components.Textbox)
+        assert isinstance(interface.input_components[0], gr.Audio)
+        assert isinstance(interface.output_components[0], gr.Textbox)
 
     def test_image_classification(self):
         model_type = "image-classification"
@@ -143,8 +143,8 @@ class TestLoadInterface:
             "models/google/vit-base-patch16-224", api_key=None, alias=model_type
         )
         assert interface.__name__ == model_type
-        assert isinstance(interface.input_components[0], gr.components.Image)
-        assert isinstance(interface.output_components[0], gr.components.Label)
+        assert isinstance(interface.input_components[0], gr.Image)
+        assert isinstance(interface.output_components[0], gr.Label)
 
     def test_feature_extraction(self):
         model_type = "feature-extraction"
@@ -154,8 +154,8 @@ class TestLoadInterface:
             alias=model_type,
         )
         assert interface.__name__ == model_type
-        assert isinstance(interface.input_components[0], gr.components.Textbox)
-        assert isinstance(interface.output_components[0], gr.components.Dataframe)
+        assert isinstance(interface.input_components[0], gr.Textbox)
+        assert isinstance(interface.output_components[0], gr.Dataframe)
 
     def test_sentence_similarity(self):
         model_type = "text-to-speech"
@@ -165,8 +165,8 @@ class TestLoadInterface:
             alias=model_type,
         )
         assert interface.__name__ == model_type
-        assert isinstance(interface.input_components[0], gr.components.Textbox)
-        assert isinstance(interface.output_components[0], gr.components.Audio)
+        assert isinstance(interface.input_components[0], gr.Textbox)
+        assert isinstance(interface.output_components[0], gr.Audio)
 
     def test_text_to_speech(self):
         model_type = "text-to-speech"
@@ -176,8 +176,8 @@ class TestLoadInterface:
             alias=model_type,
         )
         assert interface.__name__ == model_type
-        assert isinstance(interface.input_components[0], gr.components.Textbox)
-        assert isinstance(interface.output_components[0], gr.components.Audio)
+        assert isinstance(interface.input_components[0], gr.Textbox)
+        assert isinstance(interface.output_components[0], gr.Audio)
 
     def test_text_to_image(self):
         model_type = "text-to-image"
@@ -185,13 +185,14 @@ class TestLoadInterface:
             "models/osanseviero/BigGAN-deep-128", api_key=None, alias=model_type
         )
         assert interface.__name__ == model_type
-        assert isinstance(interface.input_components[0], gr.components.Textbox)
-        assert isinstance(interface.output_components[0], gr.components.Image)
+        assert isinstance(interface.input_components[0], gr.Textbox)
+        assert isinstance(interface.output_components[0], gr.Image)
 
     def test_english_to_spanish(self):
-        interface = gr.Interface.load("spaces/abidlabs/english_to_spanish")
-        assert isinstance(interface.input_components[0], gr.components.Textbox)
-        assert isinstance(interface.output_components[0], gr.components.Textbox)
+        with pytest.warns(UserWarning):
+            io = gr.Interface.load("spaces/abidlabs/english_to_spanish", title="hi")
+        assert isinstance(io.input_components[0], gr.Textbox)
+        assert isinstance(io.output_components[0], gr.Textbox)
 
     def test_sentiment_model(self):
         io = gr.Interface.load("models/distilbert-base-uncased-finetuned-sst-2-english")